se encapsula el codigo en modulos por manejo de cada tipo y se añade la opcion y funcion de arreglo de caracteres
