+++ arreglo/arreglo.h
@@ -1,0 +1,9 @@
+#ifndef ARREGLO_H
+#define ARREGLO_H
+
+// Prototipos de funciones de arreglo dinámico
+void addElement(int **array, int *size, int element); // añadir elemento al arrteglo
+void printArray(int *array, int size);                // Imprime el arreglo
+void deleteElement(int **array, int *size);
+
+#endif // ARREGLO_H
