+++ char_array/char_array.h
@@ -1,0 +1,11 @@
+#ifndef CHAR_ARRAY_H
+#define CHAR_ARRAY_H
+
+#include <stdio.h>
+#include <stdlib.h>
+
+void addElementChar(char **array, int *size, char element);
+void deleteElementChar(char **array, int *size);
+void printArrayChar(char *array, int size);
+
+#endif /* CHAR_ARRAY_H */
